Securing the backend so the user cannot delete an account via the destroy method unless they are admin.

+++ app/Http/Controllers/ProfileController.php
@@ -51,13 +51,16 @@ class ProfileController extends Controller
 
         $user = $request->user();
 
-        Auth::logout();
+        if($user['is_admin']) {
+            Auth::logout();
 
-        $user->delete();
+            $user->delete();
 
-        $request->session()->invalidate();
-        $request->session()->regenerateToken();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
 
+        }
         return Redirect::to('/');
+
     }
 }
